requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-apache-kafka>=1.9.0
apache-airflow-providers-common-messaging>=1.0.3
apache-airflow-providers-amazon
pandas>=2.3.2
pyarrow>=21.0.0
boto3>=1.40.26